destructure Result comp
